+++ src/components/Landing/components/TechCrowdCTA.tsx
@@ -12,10 +12,10 @@ export const TechCrowdCTA: React.FC<TechCrowdCTAProps> = ({ className }) => {
     <div
       className={twMerge(
         className,
-        'flex max-w-5xl justify-between gap-10 rounded-xl border border-gray-100 px-11 py-10 shadow-sm',
+        'flex max-w-5xl flex-col items-center justify-between gap-10 rounded-xl border border-gray-100 px-11 py-10 shadow-sm md:flex-row',
       )}
     >
-      <div className="flex gap-8">
+      <div className="flex flex-col items-center gap-8 md:flex-row">
         <Image
           className="max-h-28"
           src="/assets/images/tech-crowd-cta.png"
@@ -23,7 +23,7 @@ export const TechCrowdCTA: React.FC<TechCrowdCTAProps> = ({ className }) => {
           width={195}
           height={115}
         />
-        <div className="flex flex-col">
+        <div className="flex flex-col text-center md:text-start">
           <div className="text-lg font-semibold">Tech Crowd 01 (TC01)</div>
           <div className="mt-1.5 text-base text-gray-600">
             Submissions now open for all technology-focused projects in the web3
